pyproject.toml
[project]
name = "test-fmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "autogen>=0.9.1.post0",
    "fastmcp>=2.3.4",
    "google-genai>=1.15.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-core>=0.3.60",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.4.5",
    "mcp>=1.9.0",
    "mcp-flight-search>=0.2.1",
    "nest-asyncio>=1.6.0",
    "pyautogen>=0.9",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sentence-transformers>=4.1.0",
    "transformers>=4.52.3",
]
